libopkg: fix package flags on upgrade
Previously the auto_installed flag was cleared and the user flag was
set whenever a package is upgraded. This prevented upgraded dependent
packages from being removed with --autoremove.

This patch preserves the auto_installed and user flags when a package
is upgraded enabling dependent packages to be identified as such after
upgrade.

Fixes: #4
Signed-off-by: Justin Klaassen <justin@tidylabs.app>